+++ api/src/main/java/org/allaymc/api/block/component/data/BlockStateData.java
@@ -209,7 +209,7 @@ public class BlockStateData {
     }
 
     public boolean isTransparent() {
-        return translucency() != 1.0f;
+        return translucency() > 0;
     }
 
     /**