+++ CHANGELOG.md
@@ -38,6 +38,8 @@ Unless otherwise specified, any version comparison below is the comparison of se
 ### Fixed
 
 - Fixed skin display issue in 1.21.60.
+- Fixed `BlockStateData#isTransparent()` method.
+- Fixed EnderChest behaviour and drops.
 
 ### Removed